Refactor time.js
